requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
djangorestframework==3.8.2
djangorestframework-simplejwt==3.3
gunicorn==19.9.0
psycopg2-binary==2.8.5
PyJWT==1.6.4
ipdb==0.11
